@@ -37,20 +37,19 @@
   },
   "devDependencies": {
     "@changesets/cli": "^2.26.1",
-    "@types/node": "^18.15.11",
-    "@typescript-eslint/eslint-plugin": "^5.58.0",
-    "@typescript-eslint/parser": "^5.58.0",
+    "@types/node": "^18.16.16",
+    "@typescript-eslint/eslint-plugin": "^5.59.8",
+    "@typescript-eslint/parser": "^5.59.8",
     "del-cli": "^4.0.1",
-    "esbuild": "^0.17.16",
-    "eslint": "^8.38.0",
+    "esbuild": "^0.17.19",
+    "eslint": "^8.42.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-prettier": "^4.2.1",
     "npm-run-all": "^4.1.5",
-    "pnpm": "^8.2.0",
-    "prettier": "^2.8.7",
-    "sass": "^1.62.0",
-    "typescript": "^5.0.4",
-    "vite": "^4.2.1",
-    "vitest": "^0.30.1"
+    "prettier": "^2.8.8",
+    "sass": "^1.62.1",
+    "typescript": "^5.1.3",
+    "vite": "^4.3.9",
+    "vitest": "^0.31.4"
   }
 }
