@@ -12,21 +12,21 @@
     "lint": "eslint ."
   },
   "dependencies": {
-    "svelte": "^3.55.1",
+    "svelte": "^3.59.1",
     "svelte-check": "^2.10.3",
     "svelte-preprocess": "^4.10.7"
   },
   "devDependencies": {
     "@sveltejs/adapter-cloudflare": "1.1.0",
     "@sveltejs/kit": "1.3.2",
-    "@typescript-eslint/eslint-plugin": "^5.54.1",
-    "@typescript-eslint/parser": "^5.54.1",
-    "eslint": "^8.35.0",
-    "eslint-config-prettier": "^8.7.0",
+    "@typescript-eslint/eslint-plugin": "^5.59.8",
+    "@typescript-eslint/parser": "^5.59.8",
+    "eslint": "^8.42.0",
+    "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-svelte3": "^4.0.0",
-    "sass": "^1.58.3",
-    "tslib": "^2.5.0",
+    "sass": "^1.62.1",
+    "tslib": "^2.5.3",
     "typescript": "^4.9.5",
-    "vite": "^4.1.4"
+    "vite": "^4.3.9"
   }
 }
